+++ exist-core/src/main/java/org/exist/xquery/XQueryWatchDog.java
@@ -90,12 +90,12 @@ public class XQueryWatchDog {
         final Configuration conf = broker.getBrokerPool().getConfiguration();
         Object option = conf.getProperty(PROPERTY_QUERY_TIMEOUT);
         if(option != null)
-            {timeout = ((Long)option).longValue();}
+            {timeout = (Long) option;}
         if(timeout <= 0)
             {timeout = Long.MAX_VALUE;}
         option = conf.getProperty(PROPERTY_OUTPUT_SIZE_LIMIT);
         if(option != null)
-            {maxNodesLimit = ((Integer)option).intValue();}
+            {maxNodesLimit = (Integer) option;}
     }
     
     public void setTimeoutFromOption(Option option) throws XPathException {
